+++ clients/admin-ui/src/features/configure-consent/DataUsesForm.tsx
@@ -7,6 +7,7 @@ import {
   CustomCreatableSelect,
   CustomSelect,
 } from "~/features/common/form/inputs";
+import { dataUseIsConsentUse } from "~/features/configure-consent/vendor-transform";
 import { selectDataUseOptions } from "~/features/data-use/data-use.slice";
 import {
   selectDictDataUses,
@@ -15,6 +16,7 @@ import {
 import { transformDictDataUseToDeclaration } from "~/features/system/dictionary-form/helpers";
 
 import {
+  CONSENT_USE_OPTIONS,
   EMPTY_DECLARATION,
   FormValues,
   MinimalPrivacyDeclaration,
@@ -27,8 +29,15 @@ const DataUseBlock = ({
   index: number;
   isSuggestion: boolean;
 }) => {
-  const dataUseOptions = useAppSelector(selectDataUseOptions);
+  const allDataUseOptions = useAppSelector(selectDataUseOptions);
   const textColor = isSuggestion ? "complimentary.500" : "gray.800";
+
+  const { values } = useFormikContext<FormValues>();
+
+  const detailedDataUseOptions = allDataUseOptions.filter(
+    (o) =>
+      o.value.split(".")[0] === values.privacy_declarations[index].consent_use
+  );
 
   return (
     <VStack
@@ -42,10 +51,24 @@ const DataUseBlock = ({
       <CustomSelect
         label="Data use"
         tooltip="What is the system using the data for. For example, is it for third party advertising or perhaps simply providing system operations."
-        name={`privacy_declarations.${index}.data_use`}
-        options={dataUseOptions}
+        name={`privacy_declarations.${index}.consent_use`}
+        options={CONSENT_USE_OPTIONS}
         variant="stacked"
+        isRequired
+        isCustomOption
+        singleValueBlock
         textColor={textColor}
+      />
+      <CustomSelect
+        label="Detailed data use (optional)"
+        tooltip="Select a more specific data use"
+        name={`privacy_declarations.${index}.data_use`}
+        options={detailedDataUseOptions}
+        variant="stacked"
+        isCustomOption
+        singleValueBlock
+        textColor={textColor}
+        isDisabled={!values.privacy_declarations[index].consent_use}
       />
       <CustomCreatableSelect
         label="Cookie names"
@@ -71,9 +94,11 @@ const DataUsesForm = ({ showSuggestions }: { showSuggestions: boolean }) => {
   useEffect(() => {
     if (showSuggestions && values.vendor_id && dictDataUses?.length) {
       const declarations: MinimalPrivacyDeclaration[] = dictDataUses
+        .filter((du) => dataUseIsConsentUse(du.data_use))
         .map((d) => transformDictDataUseToDeclaration(d))
         .map((d) => ({
           name: d.name ?? "",
+          consent_use: d.data_use.split(".")[0],
           data_use: d.data_use,
           data_categories: d.data_categories,
           cookieNames: d.cookies?.map((c) => c.name) || [],
@@ -109,7 +134,10 @@ const DataUsesForm = ({ showSuggestions }: { showSuggestions: boolean }) => {
             disabled={
               values.privacy_declarations[
                 values.privacy_declarations.length - 1
-              ]?.data_use === EMPTY_DECLARATION.data_use
+              ]?.data_use === EMPTY_DECLARATION.data_use &&
+              values.privacy_declarations[
+                values.privacy_declarations.length - 1
+              ]?.consent_use === EMPTY_DECLARATION.consent_use
             }
             data-testid="add-data-use-btn"
           >